pipewire-pulse: Add wireplumber or pipewire-media-session-pulseaudio in Recommends (Closes: #1029763)
pipewire-pulse can be pulled by dependency without a working session
manager leading to break audio. As long as we have pipewire-media-session around,
session managers should be added in the Recommends field. Once
pipewire-media-session will be deleted, all these relations could be
reconsidered and simplified.

pipewire-pulse can work without a session manager but with very basic
features, this is why session managers are not added in the Depends field.

Signed-off-by: Dylan Aïssi <dylan.aissi@collabora.com>